requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

protobuf==3.20.0
numpy
scikit-learn
tensorflow-cpu
yfinance
pymongo